language: Make the instruction pointer a Vm struct member
There's no point in storing the stack separately but the ip only in the
run() function. I'm not sure what this bundled state will be useful for
exactly, but in theory it should allow for exiting from run() and later
calling it on the same executable again to continue execution.