language: Fix tokenization of adjacent reserved words
Input such as 'ifelse' would incorrectly get parsed as 'if' and 'else'
tokens. To fix this, the tokenizer now implements the spec's identifier
grammar as 'Identifier : IdentifierName but not ReservedWord'.
This way we can put the identifier token first in the pattern list while
still preventing it from matching reserved words.