language: Add a separate get_value bytecode instruction
For the typeof unary expression we'll need to conditionally call
GetValue on a reference - only if it is resolvable.
While there are other ways to do this, e.g. having two separate
instructions for typeof, this seems pretty nice and matches the spec
even closer :^)

Note that this is still only emitted for expressions that evaluate to
references using the existing analyze() check.