everywhere: Remove responsibility for allocator setup from Agent
Being able to pass in an allocator allows providing a customized one for
debugging purposes or an implementation other than bdwgc entirely, and
makes the global nature of its configuration less awkward in a
multi-Agent environment.

This also removes the direct dependency of zig-libgc from the kiesel
module, it's now only needed in the CLI and tests.