ci: Refactor build workflow
0a50f0a60 did not actually fix the problem (" [[: not found"), and
instead of fixing it I decided to rewrite the whole thing to not use two
separate scripts with a bunch of conditions in them - let's spell out
the full build command each time, accumulate binaries and upload them
all at once at the end.

This also fixes the issue of an inconsistent binary/version state if a
build fails half way though the workflow.