everywhere: Begin adding a platform abstraction layer
The concept of "stdout" and "stderr" doesn't work as well on more exotic
targets with a bespoke output console, and being able to override them
may also be useful for embedders in general.

As a result of this the TTY configuration for colored printing can no
longer be determined from an AnyWriter and ended up in there as well.

The stack info fits this category too as it has limited portability and
may need to be customized.