types: Stop treating Value as a tagged union
This is a first step towards introducing NaN-boxed values!

The underlying implementation remains a tagged union for now (which may
always be supported for platforms where NaN-boxing is problematic), but
its fields no longer get accessed directly. Instead, either the full
`Value.Type` can be retrieved and switched on, or individual types are
queried with `isFoo()`. Getting the underlying value is only possible
with `asFoo()`.

While I was here I flattened String and Number values into the tagged
union impl, they get wrapped on demand. Whether we can manage to put a
full slice into 64 bits somehow for String remains to be seen, but at
least for Number this is how it will work in the NaN-boxed impl.

Losing the ability to use .undefined and .null is a bit unfortunate, but
I used it as an opportunity to improve consistency a bit - everything is
a constant now:

.undefined       -> Value.undefined
.null            -> Value.null
Value.nan()      -> Value.nan
Value.infinity() -> Value.infinity