types: Make Object pointer-sized
Refactors the rest of the Object fields into Object.Data which will be useful
when we start the refactoring for NaN-boxing. We also use the allowzero
annotation on the Object.Data pointer for cleaning up the intrinsics memory
optimization.

Finally, this refactors the runtime object tag validation to only use
Object.Data.tag for validating pointers which allows us to drop two
pointer-sized values using the nifty @fieldParentPtr().