types: Store BigInt value as std.math.big.int.Managed again
This effectively reverts 3851fd548 (not cleanly as a few things changed
since then). Not that the value itself is heap-allocated we don't need
the additional savings from using the smaller Const struct, and keeping
the Managed bigint around simplifies all of the mathematical operations.

Lastly because Const stores a subslice of the allocated limbs buffer it
wasn't possible to free it properly, and testing code may not want to
rely on the GC allocator to do that for us.

I also renamed the 'value' field to 'managed' to make it clearer what it
contains.