types: Use NaN-boxing on supported platforms
The implementation requires at most 48 bits of addressable memory on the
system. Therefore, this is enabled by default on x86_64 and aarch64. It
currently supports building for x86 and other 32-bit platforms, but not
enabled by default because we should benchmark to see if the tagged
union is more performant.