language: Add labeled switch VM loop
This will give us computed gotos for our VM loop which is generally
considered to be more efficient for such hot loops. Required modifying
the Instruction enum to be exhaustive to avoid inlining the
non-exhaustive Instruction variants for instruction payloads.

Zig and/or LLVM seems to be crashing on some platforms when compiling
the labeled switch loops so we conditionally use this depending on the
target plaform.