everywhere: Pass heap-allocated types around directly as pointers
The structs with a single pointer field are a leftover from shrinking
`Value` by making the underlying types heap-allocated, cleaning this up
removes a bit of indirection and enables more optimizations - for
example Zig knows how to make `?*T` small by using a null pointer but
not (yet) `?struct{ t: *T }`.

Const correctness is somewhat sloppy but the basic idea is that
`String`, `Symbol`, and `BigInt` are always immutable and `Object`
usually is mutable. Even when there is no obvious mutation we may need
to create lazy intrinsics on first access.