language: Split up binary operator instructions
Instead of having a single instruction with one argument that we then
have to decode and switch on, just have a separate instruction for each
of them - they still all call into ApplyStringOrNumericBinaryOperator.

I've also improved the fast paths a bit while we're here: instead of
only handling i32/i32 and f64/f64, check if both are numbers, then if
both are i32, and if not fall back to coercing both to f64 so the fast
path also applies to i32/f64 and vice versa.