language: Implement typed bytecode
We now use a tagged union to represent each instruction and its payload.
Then, Executable uses comptime knowledge of instruction tags to be able
to automatically encode payloads using pointer `*align(1) T` semantics.
This does sadly mean that the bytecode could look different on different
platforms depending on Zig optimizations or the platform's endian but
I believe the benefits in the simple implementation outweigh this
consequence.