language: Add instruction to resolve a binding directly
For the common case of resolve_binding immediately followed by get_value
we can bypass reference creation entirely and store a result value
directly, which is a lot faster.

The mechanism to do this (boolean flag in the context struct) should be
generic enough to apply it to other instructions later.