execution: Rename Agent.create{Exception => ErrorObject}()
Exception and error are not quite the same thing - it just happens to be
that the spec always throws error objects, but an exception can be any
value.

This is in preparation for introducing a new Exception struct that holds
an arbitrary exception value and stack trace, so the old name becomes
ambiguous.