language: Remove ParenthesizedExpression
Judging from astexplorer.net it appears that most parsers don't have a
dedicated AST node for this, they either skip it entirely or mark
whether an expression was parenthesized or not using a boolean flag.

This makes things a bit easier as we don't need to ensure everything
sees through the additional node when doing analysis or codegen which
is error-prone.