types: Add Object.{createData,define}PropertyDirect()
Even with shapes the object property machinery is still pretty slow, so
skipping the [[GetOwnProperty]] call when we know the property doesn't
exist yet is valuable. Unfortunately this is not a speedup for array
properties yet as they still have to update the length property but I
updated those to use the new methods as well to make adding a fast path
in the future easier.