execution: Rework Agent.Platform slightly
The two GC allocators are now part of the platform implementation.
Asking users of the agent to construct one is okay, now that we have two
it's pretty awkward. As most users will use a target where libgc is
supported it makes sense to hide this as an implementation detail within
Platform.default(). For builds with libgc disabled we give you a leaking
std.heap.page_allocator instead, and if neither of those are supported
chances are you need a custom platform implementation anyway.

Additionally the caller is now always required to create the platform
itself. By moving the decision of whether Platform.default() will be
called from the agent to the caller we can make the panic on unsupported
targets a compile error.