types: Simplify Object.defineBuiltinProperty()
We already have defineBuiltinAccessor{,WithAttributes}() so the only
time you'd pass in a property descriptor is to customize the attributes.
Instead of an overly complex anytype solution we can split this into two
functions like we do for the rest of the defineBuiltin family - one with
default attributes and one where the user provides them.

This also massively cuts down on (incomplete) PropertyDescriptor use,
once defineProperty{OrThrow,Direct}() have been converted to the custom
CompletePropertyDescriptor type we could consider renaming it.