builtins: Don't migrate to sparse storage when increasing array length
Setting the length by itself should not have an effect on the underlying
storage, it only needs to be migrated when setting properties. Notably
this prevents 'new Array(N)' from being created with sparse storage, now
it will not allocate any storage until a property is set. If properties
are set in order without creating array holes dense storage will be
maintained.

This requires a small tweak to not iterate over all possible indices if
there is no storage, test262 sets an array length to 2^32-1 and back to
0 in some tests.