language: Fix incorrect line terminator precedence
The line terminators were compared one-by-one for line comments and
hashbangs, which led to incorrect behavior in a few places.

Consider this example,

// hello{CR}{LF}

The {CR} would be seen as a part of the line comment even though it is a
valid line terminator because we look for instances of {LF} first.