ci: Set Zig cache size limit of the docs job to 4 GiB
I missed this in 06ec81f26. Because docs generation is using a debug
build the bloat factor is higher, pushing the cache size just above the
2 GiB limit and causing it to be deleted after each run.

Maybe in the future we can make it so that cargo doesn't need to be
involved at all, it's just Zig autodoc after all, but for now this will
do.