builtins: Rework ArrayBuffer & friends slightly
Notably this merges the underlying ArrayBuffer and SharedArrayBuffer
object types (similar to TypedArray which is also shared for all types)
and encodes the sharability in the data block instead. This makes
accessing the underlying bytes a bit less awkward (and a bit faster).

Also embrace typed integers more - things get confusing quickly when
everything is a u53. Needs a lot of enum/int casting but I quite like
the additional type safety.