builtins: Optimize updating the Array length property even more
The saga continues! (28a4611a4, 8492b2d17)

This time we abandon the spec's way of storing the Array length property
as a regular named property and store it as a field instead, along with
its writeable attribute. This needs a few special cases in the ordinary
internal methods but is much faster and easier to work with.