Revert "types: Pass in next property index instead of storing them on the shape"
This reverts commit 762b3a7796d03f0a36bb1cd505f75490e49d5098.

After starting to experiment with pre-configured shapes for frequently
created objects it turns out that having to provide an index is awkward
when the property storage is populated later on, as usually the current
length will be the next index.

The next commit will combine value and accessor storage, so the overhead
is reduced to storing a single index per shape.