types: Allow altering non-unique shape without transition
The idea behind automatically making a shape unique once it's altered
without a transition was to avoid polluting a shape already part of a
transition chain. If we're using a fresh shape this is not an issue, so
we can relax the behavior while maintaining safety with an assertion.

This will permit pre-configured shapes to be created without any
transitions while still allowing a transition chain to exist afterwards.